Linking by shared library. Adding termalization

+++ src/molecdyn_build_ext.sh
@@ -1,14 +1,18 @@
 PYTHON=python2.7
 
 sudo rm /usr/local/lib/python2.7/dist-packages/Molec*
+sudo rm /usr/local/lib/python2.7/dist-packages/Molec*
 rm -r build
 
 make clean
-make
+make 
+#-e DVERBOSE
+if [ ! \( -L "/usr/local/lib/python2.7/dist-packages/libmolecdyn.so" \) ]
+then
+    sudo ln -s ../lib/libmolecdyn.so /usr/local/lib/python2.7/dist-packages/libmolecdyn.so
+fi
 
-sudo ln -s ../lib/libmolecdyn.so /usr/local/lib/python2.7/dist-packages/libmolecdyn.so
-
-$PYTHON --version
+echo "Installing extension for:" $PYTHON --version
 $PYTHON setup.py build
 sudo $PYTHON setup.py install
 # $PYTHON setup.py build_ext -i
